Add javadoc.

+++ src/main/java/User.java
@@ -58,6 +58,10 @@ public class User implements Serializable {
         return age;
     }
 
+    /**
+     * Increment the user activity per message.
+     * @return boolean
+     */
     public boolean incrementActivity() {
         activity++;
         long old = lastActivityTime;
@@ -71,27 +75,43 @@ public class User implements Serializable {
         return leveledUp;
     }
 
+    /**
+     * Update the level of the user.
+     * @return boolean
+     */
     private boolean updateLevel() {
         long lastLevel = level;
         level = (long) (Math.log10(exp + 1) / Math.log10(2));
         return level > lastLevel;
     }
 
+    /**
+     * Update the experience per message.
+     * @param secs
+     */
     private void updateEXP(double secs) {
         exp += calculateEXPGain(secs);
     }
 
     /**
-     * presence = activity / days
+     * Update the presence of the user.
      */
     private void updatePresence() {
         presence = (double) activity / ((double) age / Util.NANOSECONDS_PER_SECOND / 60 / 60 / 24 + 1);
     }
 
+    /**
+     * Calculate the amount of experience gain per message.
+     * @param secs The seconds since the last message.
+     * @return double
+     */
     private double calculateEXPGain(double secs) {
-        return (double) Math.min(secs / 60, 1);
+        return Math.min(secs / 60, 1);
     }
 
+    /**
+     * Update the age of the user.
+     */
     private void updateAge() {
         age = System.nanoTime() - FIRST_SEEN;
     }
